Move atn into actors and add atn commands to gws
