+++ gw_spaceheat/actors/atn/dashboard/dashboard.py
@@ -9,11 +9,11 @@ from gwproto.enums import TelemetryName
 from gwproto.named_types import PowerWatts
 
 
-from tests.atn.dashboard.misc import UpdateSources
-from tests.atn.atn_config import DashboardSettings
-from tests.atn.dashboard.channels.containers import Channels
-from tests.atn.dashboard.display.displays import Displays
-from tests.atn.dashboard.hackhp import HackHp
+from actors.atn.dashboard.misc import UpdateSources
+from actors.atn.atn_config import DashboardSettings
+from actors.atn.dashboard.channels.containers import Channels
+from actors.atn.dashboard.display.displays import Displays
+from actors.atn.dashboard.hackhp import HackHp
 from named_types import SnapshotSpaceheat
 
 class Dashboard:
