--- tests/utils/__init__.py
@@ -1,3 +1,0 @@
-from tests.utils.scada_recorder import ScadaRecorder
-
-__all__ = ["ScadaRecorder"]
