--- tests/utils/comm_events.py
@@ -1,36 +1,0 @@
-# import datetime
-# import enum
-# from dataclasses import dataclass
-# from dataclasses import field
-# from typing import List
-#
-#
-# class CommEvents(enum.Enum):
-#     invalid = "invalid"
-#     connect = "connect"
-#     connect_fail = "connect_fail"
-#     subscribe = "subscribe"
-#     message = "message"
-#     mqtt_message = "mqtt_message"
-#     publish = "publish"
-#     unsubscribe = "unsubscribe"
-#     disconnect = "disconnect"
-#     socket_open = "socket_open"
-#     socket_close = "socket_close"
-#     socket_register_write = "socket_register_write"
-#     socket_unregister_write = "socket_unregister_write"
-#
-#
-# @dataclass
-# class CommEvent:
-#     timestamp: datetime.datetime = field(default_factory=datetime.datetime)
-#     broker: str = ""
-#     event: CommEvents = CommEvents.invalid
-#     params: List = field(default_factory=list)
-#
-#     def __str__(self):
-#         param_str = str(self.params)
-#         max_param_width = 80
-#         if len(param_str) > max_param_width:
-#             param_str = param_str[:max_param_width] + "..."
-#         return f"{self.timestamp.isoformat()}  {self.broker:14s}  {self.event.value:23s}  {param_str}"
