+++ tests/utils/scada_live_test_helper.py
@@ -9,7 +9,7 @@ from gwproactor_test.tree_live_test_helper import TreeLiveTest
 from actors.config import ScadaSettings
 from data_classes.house_0_layout import House0Layout
 from tests.conftest import TEST_HARDWARE_LAYOUT_PATH
-from tests.atn.atn_app import AtnApp
+from atn_app import AtnApp
 from scada2_app import Scada2App
 from scada_app import ScadaApp
